Split arm interaction into native projects
