+++ common/src/main/java/net/woukie/createmissiles/CreateMissiles.java
@@ -38,7 +38,6 @@ public class CreateMissiles {
         MissileMenus.init();
         MissileBlocks.init();
         MissileBlockEntities.init();
-        ArmInteractionPoints.init();
     }
 
     public static CreateRegistrate registrate() {
