+++ forge/src/main/java/net/woukie/createmissiles/forge/CreateMissilesForge.java
@@ -23,6 +23,7 @@ public class CreateMissilesForge {
 
         CreateMissiles.registrate().registerEventListeners(eventBus);
         CreateMissiles.init();
+        ArmInteractionPointsForge.init();
     }
 
     private void clientSetup (final FMLClientSetupEvent event) {
